requirements.txt
streamlit==0.71.0
numpy==1.19.2
pycaret==2.2.2
pandas==1.1.2
Pillow==8.0.1
